+++ docs/changelog.md
@@ -82,6 +82,9 @@ For more information on features and bug fixes in 1.0.0, see the GitHub mileston
 * [JanusGraph zip with embedded Cassandra and ElasticSearch](https://github.com/JanusGraph/janusgraph/releases/download/v1.0.0/janusgraph-full-1.0.0.zip)
 
 #### Upgrade Instructions
+##### Drop support for HBase 1
+
+We are dropping support for HBase 1.
 
 ### Version 0.6.0 (Release Date: September 3, 2021)
 
@@ -100,7 +103,7 @@ compile "org.janusgraph:janusgraph-core:0.6.0"
 **Tested Compatibility:**
 
 * Apache Cassandra 3.0.14, 3.11.10
-* Apache HBase 1.6.0, 2.2.7
+* Apache HBase 2.2.7
 * Google Bigtable 1.3.0, 1.4.0, 1.5.0, 1.6.0, 1.7.0, 1.8.0, 1.9.0, 1.10.0, 1.11.0, 1.14.0
 * Oracle BerkeleyJE 7.5.11
 * Elasticsearch 6.0.1, 6.6.0, 7.14.0
@@ -369,7 +372,6 @@ distance, where previously they used the backend's default max distance of 2:
 | surprises | surprizes | true | true |
 | surprises | surpprises | true | true |
 | surprises | surpprisess | false | false |
-
 
 ### Version 0.5.3 (Release Date: December 24, 2020)